feat: implement modularized insights feed and lineage visualization components with state management
